This resolves the timezone issue
This resolves the timezone issue where some date/time entries miscalculated whether it was today or tomorrow.

The issue is that SugarJS doesn't take into account the timezone. This pull request forces the use of Google's `Utilities.formatDate()`, which takes into account the timezone. Previously, when when a date wasn't "converted", it simply used the SugarJS method to create the date (which didn't take into account the timezone). But with this change, the selected timezone will always be taken into account.